📦 Import `RootModel` dynamically when accessed directly only
