+++ pydantic/__init__.py
@@ -1,3 +1,5 @@
+import typing
+
 import pydantic_core
 from pydantic_core.core_schema import (
     FieldSerializationInfo,
@@ -80,7 +82,6 @@ __all__ = [
     # main
     'BaseModel',
     'create_model',
-    'RootModel',
     # network
     'AnyUrl',
     'AnyHttpUrl',
@@ -101,6 +102,8 @@ __all__ = [
     'MySQLDsn',
     'MariaDBDsn',
     'validate_email',
+    # root_model
+    'RootModel',
     # tools
     'parse_obj_as',
     'schema_of',
@@ -170,5 +173,18 @@ __all__ = [
     'GetJsonSchemaHandler',
 ]
 
+# A mapping of {<member name>: <module name>} defining dynamic imports
+__all_dynamic__ = {'RootModel': '.root_model'}
+if typing.TYPE_CHECKING:
+    from .root_model import *
 
-__getattr__ = getattr_migration(__name__)
+__getattr_migration__ = getattr_migration(__name__)
+
+
+def __getattr__(attr_name: str) -> object:
+    if attr_name in __all_dynamic__:
+        from importlib import import_module
+
+        module = import_module(__all_dynamic__[attr_name], package=__package__)
+        return getattr(module, attr_name)
+    return __getattr_migration__(attr_name)
